Squashed commit of the following:
commit 0653aa057163ce729d91748a967f1e4feaa991f1
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:36:58 2020 +0900

    Update nightly-ubuntu.yml

commit ad3039d020c6f83d16ef9f33913b77ba5b0a56d9
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:34:04 2020 +0900

    Update nightly-ubuntu.yml

commit 3838ab2a7c2edcf7e180016634a1a69f8fc1bb3a
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:30:54 2020 +0900

    Update nightly-ubuntu.yml

commit 47c989279db587eba3d1a04520c1758bef006e51
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:28:15 2020 +0900

    Update nightly-ubuntu.yml

commit a84b0056c109769eee22f3dc6e9e1fc987400d7f
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:11:00 2020 +0900

    Win-GDI 64bit build

commit 5754a76dd0834e8fef27664fa1e2d69455e3a66d
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 15:03:51 2020 +0900

    Update nightly-ubuntu.yml

commit c2b10c895db7676971eee9d3daa623100a359114
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 14:52:07 2020 +0900

    Update nightly-ubuntu.yml

commit 9574de2645d006d05349cc5fa1b6dd9dac5d93e6
Author: teamhimeh <thleaderh@gmail.com>
Date:   Sat Aug 29 13:38:20 2020 +0900

    Update nightly-ubuntu.yml

commit 6079b00904ff49929c41a067eeb110a72d0c7e6f
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 18:04:48 2020 +0900

    Update nightly-ubuntu.yml

commit 7a24c557b5ee0dab6a6feeb4723654090a4911d1
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 18:00:08 2020 +0900

    Update nightly-ubuntu.yml

commit 9f73f1e512a4e7a83651127a2134c79863b0314f
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 17:56:17 2020 +0900

    FIX: Ubuntu CI settings

commit 083b18a00d9ba7734f068b0ce804f05238504ca7
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 17:49:52 2020 +0900

    Ubuntu CI settings

commit f01f4770979340c80ceddc9e9f13bcba41fad3ab
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 11:01:33 2020 +0900

    FIX: macOS CI settings

commit 0fd87ce4d15591cd4468df6d563e2de5f46b27c5
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 10:54:00 2020 +0900

    macOS CI settings

commit 3b7ef71a0ca46a6c40921854cc3c0c7e9fc48476
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 01:02:12 2020 +0900

    32bit mingw CI settings

commit 3875bd5a0c47a912297781a0d7c2bca2515ae654
Merge: 0bf727c44 168eaf2eb
Author: teamhimeh <thleaderh@gmail.com>
Date:   Fri Aug 28 00:07:39 2020 +0900

    Merge branch 'master' into OTRP-CI

commit 168eaf2eb1e5b068071e1bc252ee6884e0f203b9
Merge: 83fc8f113 dcb84a535
Author: teamhimeh <thleaderh@gmail.com>
Date:   Thu Aug 27 17:55:05 2020 +0900

    r9186 Merge remote-tracking branch 'upstream/master'

commit dcb84a535335cafe0cb3d54320f0a15a41eaa21c
Author: Markus Pristovsek <prissi@physik.tu-berlin.de>
Date:   Mon Aug 24 02:54:16 2020 +0000

    FIX: typo, SDL2 instead freetype

    git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9186 8aca7d54-2c30-db11-9de9-000461428c89

commit 6cb3406c1c34650c61c6722dabc0c7cbd9dba43e
Author: Markus Pristovsek <prissi@physik.tu-berlin.de>
Date:   Sun Aug 23 14:47:13 2020 +0000

    FIX: nighly build on github also for SDL2 64bit

    git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9185 8aca7d54-2c30-db11-9de9-000461428c89

commit faad35e8bfbbd6d0fc1dbdef6aa8069dc2c81f6a
Author: Markus Pristovsek <prissi@physik.tu-berlin.de>
Date:   Sat Aug 22 13:52:13 2020 +0000

    ADD: github action for 32bit GDI build as well

    git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@9184 8aca7d54-2c30-db11-9de9-000461428c89

+++ vehicle/simvehicle.cc
@@ -5148,7 +5148,7 @@ air_vehicle_t::~air_vehicle_t()
 void air_vehicle_t::set_convoi(convoi_t *c)
 {
 	DBG_MESSAGE("aircraft_t::set_convoi()","%p",c);
-	if(leading  &&  (unsigned long)cnv > 1) {
+	if(leading  &&  (uintptr_t)cnv > 1) {
 		// free stop reservation
 		route_t const& r = *cnv->get_route();
 		if(target_halt.is_bound()) {
